azure-functions-backend/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure Functions
azure-functions>=1.15.0

# Data Processing
yfinance==0.2.54
pandas==2.2.0
numpy==1.26.4

# AI and ML
openai==1.60.1

# HTTP and API
requests==2.32.3

# News API
newsapi-python==0.2.7

# Added from the code block
fastapi==0.115.7
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.2
typing-extensions==4.8.0
numpy==1.24.3
pandas==2.0.3